fix(color-picker): providing own color-collection service per color-picker
