+++ src/lib/color-picker/color-picker-collection.service.ts
@@ -1,0 +1,38 @@
+import { Inject, Injectable } from '@angular/core';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { COLOR_STRING_FORMAT, ColorFormat, formatColor, parseColorString } from './color-picker';
+
+
+@Injectable()
+export class MccColorPickerCollectionService {
+
+  /**
+   * Hold current selected color
+   */
+  private _selectedColor: BehaviorSubject<string> = new BehaviorSubject<string>('');
+
+
+  constructor(@Inject(COLOR_STRING_FORMAT) private colorStringFormat: ColorFormat) {
+  }
+
+  /**
+   * Change internal selected color
+   */
+  changeSelectedColor(colorString: string) {
+    const color = parseColorString(colorString);
+
+    if (!colorString || !color) {
+      return;
+    }
+
+    const clrString = formatColor(color, this.colorStringFormat);
+    this._selectedColor.next(clrString);
+  }
+
+  /**
+   * Return internal selected color
+   */
+  getSelectedColor(): Observable<string> {
+    return this._selectedColor.asObservable();
+  }
+}
