+++ src/lib/color-picker/color-picker.module.ts
@@ -19,13 +19,12 @@ import {
   USED_COLORS
 } from './color-picker';
 
-import { MccColorPickerService } from './color-picker.service';
-
 import { MccColorPickerComponent } from './color-picker.component';
 import { MccColorPickerSelectorComponent } from './color-picker-selector.component';
 import { MccColorPickerCollectionComponent } from './color-picker-collection.component';
 import { MccColorPickerOriginDirective, MccConnectedColorPickerDirective, } from './color-picker-origin.directive';
 import { MccColorPickerOptionDirective } from './color-picker-option.directive';
+import { MccColorPickerService } from './color-picker.service';
 
 @NgModule({
   imports: [
